made mechanics part of gwent game as solely dependent upon it and had circular dependancy issue
