+++ gwentSim.pro
@@ -15,7 +15,6 @@ TEMPLATE = app
 SOURCES += main.cpp\
         mainwindow.cpp \
     gwentGame.cpp \
-    gwentMechanics.cpp \
     gwentCard.cpp \
     gwentPlayer.cpp \
     gwentBoard.cpp \
@@ -23,7 +22,6 @@ SOURCES += main.cpp\
 
 HEADERS  += mainwindow.h \
     gwentGame.h \
-    gwentMechanics.h \
     gwentPlayer.h \
     gwentCard.h \
     gwentBoard.h \
